+++ TDEngine2/include/ecs/CComponentManager.h
@@ -159,7 +159,7 @@ namespace TDEngine2
 			TDE2_API static E_RESULT_CODE RegisterUniqueComponentType(TypeId typeId);
 
 #if TDE2_EDITORS_ENABLED
-			TDE2_API const std::vector<std::string>& GetRegisteredComponentsIdentifiers() const override;
+			TDE2_API const std::vector<TComponentTypeInfo>& GetRegisteredComponentsIdentifiers() const override;
 #endif
 		protected:
 			DECLARE_INTERFACE_IMPL_PROTECTED_MEMBERS(CComponentManager)
@@ -207,7 +207,7 @@ namespace TDEngine2
 			TUniqueComponentsTable   mUniqueComponentsRegistry;
 
 #if TDE2_EDITORS_ENABLED
-			std::vector<std::string> mRegisteredComponentTypes;
+			std::vector<TComponentTypeInfo> mRegisteredComponentTypes;
 #endif
 	};