+++ TDEngine2/include/ecs/IComponentManager.h
@@ -266,7 +266,7 @@ namespace TDEngine2
 			}
 
 #if TDE2_EDITORS_ENABLED
-			TDE2_API virtual const std::vector<std::string>& GetRegisteredComponentsIdentifiers() const = 0;
+			TDE2_API virtual const std::vector<TComponentTypeInfo>& GetRegisteredComponentsIdentifiers() const = 0;
 #endif
 		protected:
 			DECLARE_INTERFACE_PROTECTED_MEMBERS(IComponentManager)